Fix Blazor Client build errors: types, casting and local storage
- Fixed `decimal` to `float` conversion error in `UtilsTamanhoServices` by using `decimal`.
- Updated `UserSessionService`, `CarrinhoComprasTemporarioService`, and `FavoritosTemporariosService` to use correct `Blazored.LocalStorage` methods (`GetItemAsync<string>`/`SetItemAsync`).
- Fixed potential `NullReferenceException` in `SlideProdutoComponent` image handling.
- Ensured consistent typing across client services.

+++ src/MyMEDIA/MyMEDIA.Client/Services/Interfaces/IUtilsTamanhoServices.cs
@@ -3,7 +3,7 @@
     public interface IUtilsTamanhoServices
     {
         int Index { get; set; }
-        float PrecoRefSize { get; set; }
+        decimal PrecoRefSize { get; set; }
 
         public event Action OnChange;
     }
